CRM权限管理

+++ src/main/java/com/situ/scrm/sys/area/dao/AreaDao.java
@@ -21,5 +21,5 @@ import com.situ.scrm.sys.area.domain.Area;
 @Repository
 public interface AreaDao {
 	
-	List<Area> findByParent(Long parentCode);
+	List<Area> findByParent(Integer parentCode);
 }
